KAFKA-14485: Move LogCleaner exceptions to storage module
