+++ storage/src/main/java/org/apache/kafka/storage/internals/log/ThreadShutdownException.java
@@ -1,10 +1,10 @@
-/**
+/*
  * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
+ * contributor license agreements. See the NOTICE file distributed with
  * this work for additional information regarding copyright ownership.
  * The ASF licenses this file to You under the Apache License, Version 2.0
  * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
+ * the License. You may obtain a copy of the License at
  *
  *    http://www.apache.org/licenses/LICENSE-2.0
  *
@@ -14,11 +14,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
-package kafka.common
+package org.apache.kafka.storage.internals.log;
 
 /**
  * An exception that indicates a thread is being shut down normally.
  */
-class ThreadShutdownException extends RuntimeException {
+public class ThreadShutdownException extends RuntimeException {
 }
